+++ src/detect-flowvar.h
@@ -28,8 +28,8 @@ typedef struct DetectFlowvarData_ {
     char *name;
     uint32_t idx;
     uint8_t *content;
-    uint8_t content_len;
-    uint8_t flags;
+    uint16_t content_len;
+    uint32_t flags;
 } DetectFlowvarData;
 
 /* prototypes */
